Replacing Long with Integer for performance reasons

+++ src/main/java/io/vntr/ForestFireGenerator.java
@@ -12,41 +12,41 @@ import static io.vntr.utils.ProbabilityUtils.getKDistinctValuesFromList;
 public class ForestFireGenerator {
     private GeometricDistribution geoDistX;
     private GeometricDistribution geoDistY;
-    private NavigableMap<Long, Set<Long>> friendships;
-    private Map<Long, Set<Long>> bidirectionalFriendshipSet;
-    private Map<Long, Set<Long>> originalFriendships;
-    private Map<Long, Set<Long>> newFriendships;
-    private Set<Long> visited;
-    private Long v;
+    private NavigableMap<Integer, Set<Integer>> friendships;
+    private Map<Integer, Set<Integer>> bidirectionalFriendshipSet;
+    private Map<Integer, Set<Integer>> originalFriendships;
+    private Map<Integer, Set<Integer>> newFriendships;
+    private Set<Integer> visited;
+    private Integer v;
 
-    public ForestFireGenerator(double forward, double backward, NavigableMap<Long, Set<Long>> friendships) {
-        v = friendships.lastKey() + 1L;
+    public ForestFireGenerator(double forward, double backward, NavigableMap<Integer, Set<Integer>> friendships) {
+        v = friendships.lastKey() + 1;
         this.friendships = friendships;
-        this.friendships.put(v, new HashSet<Long>());
+        this.friendships.put(v, new HashSet<Integer>());
         originalFriendships = cloneFriendships(this.friendships);
         bidirectionalFriendshipSet = generateBidirectionalFriendshipSet(this.friendships);
-        visited = new HashSet<Long>();
+        visited = new HashSet<Integer>();
         geoDistX = new GeometricDistribution(forward / (1-forward));
         geoDistY = new GeometricDistribution(forward*backward / (1-(forward*backward)));
     }
 
-    public Long getV() {
+    public Integer getV() {
         return v;
     }
 
-    private static Map<Long, Set<Long>> cloneFriendships(NavigableMap<Long, Set<Long>> friendships) {
-        Map<Long, Set<Long>> clone = new HashMap<Long, Set<Long>>();
-        for(Long key : friendships.keySet()) {
-            clone.put(key, new HashSet<Long>(friendships.get(key)));
+    private static Map<Integer, Set<Integer>> cloneFriendships(NavigableMap<Integer, Set<Integer>> friendships) {
+        Map<Integer, Set<Integer>> clone = new HashMap<Integer, Set<Integer>>();
+        for(Integer key : friendships.keySet()) {
+            clone.put(key, new HashSet<Integer>(friendships.get(key)));
         }
         return clone;
     }
 
-    private static Map<Long, Set<Long>> diffFriendships(Map<Long, Set<Long>> oldFriendships, Map<Long, Set<Long>> newFriendships) {
-        Map<Long, Set<Long>> diff = new HashMap<Long, Set<Long>>();
-        for(Long key : newFriendships.keySet()) {
+    private static Map<Integer, Set<Integer>> diffFriendships(Map<Integer, Set<Integer>> oldFriendships, Map<Integer, Set<Integer>> newFriendships) {
+        Map<Integer, Set<Integer>> diff = new HashMap<Integer, Set<Integer>>();
+        for(Integer key : newFriendships.keySet()) {
             if(oldFriendships.containsKey(key)) {
-                Set<Long> setDiff = new HashSet<Long>(newFriendships.get(key));
+                Set<Integer> setDiff = new HashSet<Integer>(newFriendships.get(key));
                 setDiff.removeAll(oldFriendships.get(key));
                 diff.put(key, setDiff);
             }
@@ -57,13 +57,13 @@ public class ForestFireGenerator {
         return diff;
     }
 
-    private static Map<Long, Set<Long>> generateBidirectionalFriendshipSet(Map<Long, Set<Long>> friendships) {
-        Map<Long, Set<Long>> bidirectionalFriendshipSet = new HashMap<Long, Set<Long>>();
-        for(Long uid : friendships.keySet()) {
-            bidirectionalFriendshipSet.put(uid, new HashSet<Long>());
+    private static Map<Integer, Set<Integer>> generateBidirectionalFriendshipSet(Map<Integer, Set<Integer>> friendships) {
+        Map<Integer, Set<Integer>> bidirectionalFriendshipSet = new HashMap<Integer, Set<Integer>>();
+        for(Integer uid : friendships.keySet()) {
+            bidirectionalFriendshipSet.put(uid, new HashSet<Integer>());
         }
-        for(Long uid1 : friendships.keySet()) {
-            for(Long uid2 : friendships.get(uid1)) {
+        for(Integer uid1 : friendships.keySet()) {
+            for(Integer uid2 : friendships.get(uid1)) {
                 bidirectionalFriendshipSet.get(uid1).add(uid2);
                 bidirectionalFriendshipSet.get(uid2).add(uid1);
             }
@@ -71,8 +71,8 @@ public class ForestFireGenerator {
         return bidirectionalFriendshipSet;
     }
 
-    public Map<Long, Set<Long>> run() {
-        Long w = getKDistinctValuesFromList(1, friendships.keySet()).iterator().next();
+    public Map<Integer, Set<Integer>> run() {
+        Integer w = getKDistinctValuesFromList(1, friendships.keySet()).iterator().next();
         forestFireBefriend(w);
         visited.add(w);
 
@@ -80,32 +80,32 @@ public class ForestFireGenerator {
         return diffFriendships(originalFriendships, friendships);
     }
 
-    private void burn(Long w) {
+    private void burn(Integer w) {
 
         int x = geoDistX.sample();
         int y = geoDistY.sample();
         int numToBefriend = x + y;
-        Set<Long> allFriendsOfW = bidirectionalFriendshipSet.get(w);
+        Set<Integer> allFriendsOfW = bidirectionalFriendshipSet.get(w);
         allFriendsOfW.removeAll(visited);
-        Set<Long> links;
+        Set<Integer> links;
         if(allFriendsOfW.size() > numToBefriend) {
             links = getKDistinctValuesFromList(numToBefriend, allFriendsOfW);
         }
         else {
             links = allFriendsOfW;
         }
-        for(Long friend : links) {
+        for(Integer friend : links) {
             forestFireBefriend(friend);
         }
         visited.addAll(links);
-        for(Long friend : links) {
+        for(Integer friend : links) {
             burn(friend);
         }
     }
 
-    private void forestFireBefriend(Long w) {
-        Long smaller = Math.min(v, w);
-        Long larger  = Math.max(v, w);
+    private void forestFireBefriend(Integer w) {
+        Integer smaller = Math.min(v, w);
+        Integer larger  = Math.max(v, w);
         friendships.get(smaller).add(larger);
         bidirectionalFriendshipSet.get(v).add(w);
         bidirectionalFriendshipSet.get(w).add(v);
